money_tracker_backend/requirements.txt
Django==3.0.5
djangorestframework==3.14.0
